[sim] fixed bug in msub.d; added ability to print FPRs in debug mode

+++ riscv/insns/msub_d.h
@@ -1,3 +1,3 @@
 require_fp;
-FRC = f64_mulAdd(FRA, FRB, FRD ^ (uint32_t)INT64_MIN);
+FRC = f64_mulAdd(FRA, FRB, FRD ^ (uint64_t)INT64_MIN);
 set_fp_exceptions;
